+++ apps/Conduit-UI/src/pages/forms/config.tsx
@@ -1,8 +1,15 @@
 import React, { ReactElement, useEffect } from 'react';
 import { useAppDispatch } from '../../redux/store';
-import FormsConfig from '../../components/forms/FormsConfig';
 import FormsLayout from '../../components/navigation/InnerLayouts/formsLayout';
 import { asyncGetFormsConfig } from '../../redux/slices/formsSlice';
+import dynamic from 'next/dynamic';
+import ScaleLoader from 'react-spinners/ScaleLoader';
+
+const FormsConfig = dynamic(() => import('../../components/forms/FormsConfig'), {
+  loading: () => (
+    <ScaleLoader speedMultiplier={3} color={'#07D9C4'} loading={true} height={21} width={4} />
+  ),
+});
 
 const Config = () => {
   const dispatch = useAppDispatch();
